requirements-dev.txt
python-telegram-bot==3.3
pytest==2.9.1